backend/requirements.txt
Flask
flask-cors
requests
google-generativeai
python-dotenv
gunicorn
